% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selic.R
\encoding{UTF-8}
\name{get_selic}
\alias{get_selic}
\title{Get Selic market expectations data (Focus/BCB)}
\usage{
get_selic(
  indicator = "Selic",
  first_date = Sys.Date() - 2 * 365,
  last_date = Sys.Date(),
  meeting = NULL,
  be_quiet = FALSE,
  use_memoise = FALSE,
  do_parallel = FALSE
)
}
\arguments{
\item{indicator}{A character vector with economic indicators names, see details for possible value. Defaults to "Selic".}

\item{first_date}{A single string or Date in "YYYY-mm-dd" format. Default for 2 years prior to the current date.}

\item{last_date}{A single string or Date in "YYYY-mm-dd" format. Default is the current date.}

\item{meeting}{A single string in "RX/YYYY" format, where X is a positive integer, indicating the meeting for which the statistic is expected. Defaults to NULL.}

\item{be_quiet}{Logical. Should messages or warnings not be displayed? Defaults to FALSE.}

\item{use_memoise}{Logical. Sets the use of caching system, creating a folder named "cache_bcb" by default. Defaults to FALSE.}

\item{do_parallel}{Logical. For using parallel data importation. Defaults to FALSE.}
}
\value{
A tibble with data from BCB-Olinda Data Services.
}
\description{
This function provides the extraction of data and statistics related to expectations of economic indicators, specifically Selic market expectations, made available by the Central Bank of Brazil's Expectations System through the Olinda API. The data comes from several financial institutions: banks, funds, research houses, etc. Important: arguments are case sensitive.
}
\details{
For periods for which there are no statistics, they will be omitted from the query.

Possible values for indicator argument: 'Selic'.
}
\examples{
df1 <- get_selic(indicator = "Selic", first_date = Sys.Date() - 30)
df1

df2 <- get_selic(indicator = "Selic", first_date = NULL, meeting = "R1/2024")
df2
}
\author{
Fernando da Silva <\href{mailto:fernando@fortietwo.com}{fernando@fortietwo.com}>
}
